Fix/casting orders status (#3199)
* cast order status in payment modules

* cast order status in payment modules

+++ includes/modules/payment/cod.php
@@ -49,7 +49,7 @@ class cod {
     if ($this->check() > 0) {
       $this->limit_subtotal = (double)MODULE_PAYMENT_COD_LIMIT_ALLOWED;
       if ((int) MODULE_PAYMENT_COD_ORDER_STATUS_ID > 0) {
-        $this->order_status = MODULE_PAYMENT_COD_ORDER_STATUS_ID;
+        $this->order_status = (int) MODULE_PAYMENT_COD_ORDER_STATUS_ID;
       }
     }
